[Producer] More granular exceptions (reasons below)
The point of this is that these exceptions come from different places,
and it is not the case that in the course of handling one, you should be
expected to handle all the others. A common mistake would be to handle
one of them, and then get a pattern match failure because it was a
different one of the same type that got thrown.