[Queue] implement takeQueueTimeout w/o bit atomically block
The previous version had the whole thing inside an `atomically` block,
which resulted in horrible performance, since the entire operation would
get retried all the time. This version worked sort of OK with `TQueue`,
but it could result in the system locking up for many seconds with
`TChan`.

This time, I have implemented it such that only the individual reads
from the queue are atomic; it should be functionally equivalent, if I am
not mistaken.